Playerprocessinfo improvements
I made some improvement in confluence skin, updated some button in video osd, for better resolution.

Added to auto hide video osd , in the skin settings. 

Added playerprocessinfo button to the video osd.

Added the button OSDHelpText for video osdbuttons.

Added and some values, and  reorganized the playerprocessinfo, to fit for kodi, and for coreelec.